+++ esperanto-tools-libs/src/server/Server.h
@@ -10,16 +10,21 @@
 #pragma once
 #include "runtime/IProfiler.h"
 #include "runtime/IRuntime.h"
+
 #include <device-layer/IDeviceLayer.h>
 #include <hostUtils/threadPool/ThreadPool.h>
+
+#include <memory>
 #include <string_view>
 #include <thread>
 #include <vector>
+
 namespace rt {
 class Worker;
 class Server {
 public:
-  explicit Server(const std::string& socketPath, std::unique_ptr<dev::IDeviceLayer> deviceLayer, Options options);
+  explicit Server(const std::string& socketPath, std::shared_ptr<dev::IDeviceLayer> const& deviceLayer,
+                  Options options);
   ~Server();
   void removeWorker(Worker* worker);
   void setProfiler(std::unique_ptr<profiling::IProfilerRecorder>&& profiler) {
@@ -40,7 +45,7 @@ private:
   bool running_ = true;
   std::mutex mutex_;
   std::thread listener_;
-  std::unique_ptr<dev::IDeviceLayer> deviceLayer_;
+  std::shared_ptr<dev::IDeviceLayer> deviceLayer_;
   std::unique_ptr<IRuntime> runtime_;
   std::vector<std::unique_ptr<Worker>> workers_;
   threadPool::ThreadPool tp_{1}; // used to remove workers